+++ NEWS.md
@@ -9,6 +9,7 @@ BUGFIXES :
 * `setSimulationPathAPI()` : encode URL before reading the data in simulation mode
 * `api_get()` : add warn_for_status in importFrom section
 * `readAntares()` : In disk mode, return all the available columns for a short-term storage output and match the column with the content
+* `.importOutput()` : check if output file exists in API mode (`.check_output_files_existence()`)
 
 BREAKING CHANGES :